sinks/influxdb - move batch-related logic to batch.go
